requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
importlib-metadata==4.12.0
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
